Implement the mempool interface in pure style
The `Ouroboros.Consensus.Mempool.Impl` module was split in `Impl`, `Impl.Pure`
and `Impl.Types` where:
- `Impl`: contains the monadic side of the implementation, bridging between
  operations on an `InternalState` and the actual `Mempool` API.
- `Impl.Pure`: contains the pure side of the implementation, functions that
  update an `InternalState` according to the rules described on the `Mempool`
  API. Also contains functions to "run" those pure computations on the monadic
  context to bridge with the `Impl` module.
- `Impl.Types`: define the `InternalState` used to fulfill the requirements of
  the implementation, and the `ValidationResult` used to extend the mempool. It
  is used by both `Impl` and `Impl.Pure`.

From outside of this module, the change should be transparent, as other modules
should only access the mempool via the API, and all relevant items are
re-exported through `Ouroboros.Consensus.Mempool`.
