+++ ouroboros-consensus/src/Ouroboros/Consensus/Mempool.hs
@@ -1,4 +1,6 @@
+-- | The mempool API and implementation.
 module Ouroboros.Consensus.Mempool (module X) where
 
 import           Ouroboros.Consensus.Mempool.API as X
 import           Ouroboros.Consensus.Mempool.Impl as X
+import           Ouroboros.Consensus.Mempool.TxSeq as X (TicketNo)
